+++ lib/sprockets/loader.rb
@@ -26,8 +26,13 @@ module Sprockets
     def load(uri)
       filename, params = parse_asset_uri(uri)
       if params.key?(:id)
-        asset = cache.fetch("asset-uri:#{VERSION}#{uri}") do
-          load_asset_by_id_uri(uri, filename, params)
+        unless asset = cache.get("asset-uri:#{VERSION}:#{uri}", true)
+          id = params.delete(:id)
+          uri_without_id = build_asset_uri(filename, params)
+          asset = load_asset_by_uri(uri_without_id, filename, params)
+          if asset[:id] != id
+            @logger.warn "Sprockets load error: Tried to find #{uri}, but latest was id #{asset[:id]}"
+          end
         end
       else
         asset = fetch_asset_from_dependency_cache(uri, filename) do |paths|
@@ -45,28 +50,7 @@ module Sprockets
     end
 
     private
-      def load_asset_by_id_uri(uri, filename, params)
-        # Internal assertion, should be routed through load_asset_by_uri
-        unless id = params.delete(:id)
-          raise ArgumentError, "expected uri to have an id: #{uri}"
-        end
-
-        uri = build_asset_uri(filename, params)
-        asset = load_asset_by_uri(uri, filename, params)
-
-        if id && asset[:id] != id
-          raise VersionNotFound, "could not find specified id: #{uri}##{id}"
-        end
-
-        asset
-      end
-
       def load_asset_by_uri(uri, filename, params)
-        # Internal assertion, should be routed through load_asset_by_id_uri
-        if params.key?(:id)
-          raise ArgumentError, "expected uri to have no id: #{uri}"
-        end
-
         unless file?(filename)
           raise FileNotFound, "could not find file: #{filename}"
         end